+++ src/test/completion.test.ts
@@ -8,6 +8,7 @@ describe('Should do completion in Camel URi after "ti"', () => {
 	const docUriXml = getDocUri('apacheCamel.xml');
 	const docUriJava = getDocUri('apacheCamel.java');
 	const docUriGroovy = getDocUri('helloworld.camelk.groovy');
+	const docUriCamelKafkaConnectProperties = getDocUri('camelKafkaConnect.properties');
 	const expectedCompletion = [
 		{ label: 'tika:operation'},
 		{ label: 'timer:timerName'}
@@ -27,6 +28,12 @@ describe('Should do completion in Camel URi after "ti"', () => {
 
 	it('Completes components for Groovy', async () => {
 		await testCompletion(docUriGroovy, new vscode.Position(0, 8), {
+			items: expectedCompletion
+		});
+	});
+
+	it('Completes components for Camel Kafka Connect properties', async () => {
+		await testCompletion(docUriCamelKafkaConnectProperties, new vscode.Position(7, 19), {
 			items: expectedCompletion
 		});
 	});