+++ Changelog.md
@@ -8,6 +8,8 @@
     - filename _*.groovy_ and starting with _// camel-k:_
     - filename _*.groovy_ and starting with _#!/usr/bin/env camel-k_ (not yet supported by Camel-K runtime see [here](https://github.com/apache/camel-k/issues/754))
   - Please note that all Camel components are provided in completion, specific Camel-K component are not filtered
+- Provide completion for Camel URI in Camel Kafka Connect properties file
+  - properties file needs to use '=' without spaces notation
 - Upgrade from Camel 2.24.0 to 2.24.1
 
 ## 0.0.15